backend/pyproject.toml
[project]
name = "yi-hot-monitor-backend"
version = "0.1.0"
description = "AI Hotspot Monitor Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "aiomysql>=0.2.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "python-socketio>=5.0.0",
    "apscheduler>=3.10.0",
    "httpx>=0.27.0",
    "selectolax>=0.3.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langgraph>=0.2.0",
    "aiosmtplib>=3.0.0",
    "cryptography>=46.0.7",
    "alembic>=1.18.4",
    "alembic-utils>=0.8.8",
    "pymysql>=1.1.2",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.hatch.build.targets.editable]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
